Support 1.363 GHz for cpu clocks of MSM8916 v2.0, fix throttle
8916 v2.0 supports a FMAX of 1.3632 GHz. Modify the CPR
configuration to support this frequency. Use a
combination of speed-bin and pvs-version to identity
the various FMAX parts.

Speed-bin differentiates between 1.2 GHz and 1.3632 GHz
FMAX parts and pvs-version is used to identify the
1.15 GHz parts. The combination of efuse1 and efuse
register will be used to determine the speed-bin and version to select
the correct fmax of the parts. Don't throttle at 998mhz, but 1094mhz, 998mhz not needed so delete it