ARM: 8158/1: LLVMLinux: use static inline in ARM ftrace.h
With compilers which follow the c99 standard (like modern versions of gcc and clang),"extern inline" does the wrong thing (emits code for an externaly linkable version of the inline function). In this case using static inline and removing the NULL version if return_address in return_address.c does the right thing.